# src/main/java/com/banking/onlinebanking/dto/TransactionDetailsRequestDto.java
package com.banking.onlinebanking.dto;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import lombok.Data;

@Data
public class TransactionDetailsRequestDto {
    @NotBlank(message = "account number required")
    @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8}$",message = "enter a valid account number")
    private String accountNumber;
    @Min(value = 1, message = "transaction amount should be greater than 1")
    private Double amount;
}